rebase new test

+++ src/v0.8/interfaces/core/IFilplus.sol
@@ -22,7 +22,7 @@ pragma solidity ^0.8.21;
 interface IFilplus {
     // Public getter function to access datasetRuleMaxReplicasInCountries
     function getDatasetRuleMaxReplicasInCountry(
-        uint64 _countryCode
+        uint32 _countryCode
     ) external view returns (uint16);
 
     // Set functions for public variables
@@ -35,7 +35,7 @@ interface IFilplus {
     ) external;
 
     function setDatasetRuleMaxReplicasInCountry(
-        uint64 _countryCode,
+        uint32 _countryCode,
         uint16 _newValue
     ) external;
